Add an empty theme skeleton
This is just a copy of the "simple" theme just with a favicon set.
This commit also outlines a directory structure for some theme related
files (CSS, fonts, images).
The README.md in there will be copied over verbatim to the output. I
don't think that it's a big problem. We can still delete them once we
have actual files in those folders. For now we should keep them as empty
folders can't be checked into git.