Start to parse old posts
All posts that were considered "news" had a 'category' of "news"

If they didn't have a category "news", consider them blog posts.

We are using directories as category types in Pelican.  So there
are two directories for:

1. content/news/ for 'news' items.
2. content/blog/ for 'blog' posts.

In the direcotry fix-old-posts/ all of the posts that were
categorized as 'news' now reside in the sub-dir fix-old-posts/news/

Still need to go through and clean up frontmatter:

* Change 'categories:' frontmatter to 'tags:'
* Merge with 'tags' if it already exists
* Remove 'blog' or 'news' from tags, or leave these tags in
  if we want.  I don't think they hurt anything staying except
  that there will now also be a "tag" for 'blog' and 'news'