Add script that adds legacy lede to .md files
This script reads a CSV file (separated by ';') of worpdress_id->image
mappings and ammends all .md files in content/ accordingly.

The CSV file isn't part of this commit as it's 16MB with the images
together.

TODO: Decide where to actually put the images and adapt the script to
it.