Add plugin for handling wordpress_lede images
Added `attach_ledeimg` plugin.

The `wordpress_lede` and eventually `lede` metadata point to a
local img file to use as a lede image.  Unfortunately, the Pelican
build system doesn't pickup metadata to use with {attach} or
{filename}.

({attach} allows Pelican to place files right next to the calling
files - so if you write <img src='{attach}image.jpg'> in your
markdown file, Pelican will move the `image.jpg` file to be next
to your output file as well).

This plugin takes advantage of this fact by adding a tag to the
article before it gets written:

    <img src='{attach}LEDE_IMG.jpg' class='hidden-lede'>

Which triggers the Pelican engine to place the image alongside
the output file.  Then, before the article generator finishes up
we capture one more signal to go and remove the tag we just wrote
so it doesn't show up in the final HTML.