Add plugin rm_nbsp_title to fix typogrify widont on titles
Typogrify adds an &nbsp; between the last two words in a title
to keep the last word from becoming an orphan.  This is fine
normally except when looking at titles on mobile that use long
words.  It breaks formatting on mobile. :(

This commit adds a plugin that reverts the widont behavior on
titles by removing any &nbsp; from titles.

Doing it this way still allows typogrify to do neat things with
titles otherwise (smart quotes, etc...).