Use thumbnails for small images and ledes
Using the scaled down versions for ledes in the overview pages brings
the page size down quite a lot, in a random test it was 2.4 mb -> 800k.

We now call the thumbnails just <filename>_thumbs.<ext>, so the
thumbnail size is no longer part of the name. Therefore cleaning output/
would make sense to get rid of stale files.