Fix nav being hidden, style lede-author text
When making the lede-author parent position: relative, it kicked
on the browser z-index, which basically just made the lede sit
over the navigation bar.  I simply added a z-index: 1; to the nav
menu bar, which fixed things right up.

I also styled @houz's lede-author text to be white on a black
text-shadow background to provide enough contrast to be legible
even with different backgrounds.