Fix search page title styling and add width/height to page ledes
The search page title styling was getting borked by the recent
changes to move titles up, this commit fixes that.

The page ledes didn't have an explicit width/height set, so we
got a FOUC (Flash of Unstyled Content) until the image gets
fetched, so this commit also adds a width/height attribute
explicitly to page ledes in the template (960x402).  This fixes
the FOUC when clicking through or hard-refreshing the page.