Colour -> color
The article used sometimes colour (UK), sometimes color (US). The
majority seems to be on the US side so let's adopt it for all
instances.

We still have "grey" (UK) everywhere and not "gray" (US), so the
article is not really US-Vs-UK consistant, but at least we are for the
word "color" which is probably the most important one for an article
about photography ;-).