Describe additional Mac installation error
The error "macOS cannot verify that this app is free from malware" happens on MacOS 15.7.2 when installing a nightly build. (I did not test an official release.) This error deserves a separate section since the usual workaround (open the app via the context menu) does not give any option to bypass the warning, so we shouldn't write that as a solution.